back to green

+++ app/Jobs/ProcessBookingDeleted.php
@@ -11,7 +11,7 @@ use Illuminate\Foundation\Queue\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Bus\Dispatchable;
 
-class ProcessBookingDeleted implements ShouldQueue
+class ProcessBookingDeleted
 {
     use Dispatchable, InteractsWithQueue, Queueable, SerializesModels;
 
@@ -20,12 +20,18 @@ class ProcessBookingDeleted implements ShouldQueue
     public function __construct(Bookings $booking)
     {
         $this->booking = $booking;
+        Log::info('Processing booking deletion for booking ID: ' . $booking->id);
     }
 
     public function handle()
     {
         try {
-            $calendarService = new CalendarService($this->booking->band);
+            if ($this->booking->band->bookingCalendar === null) {
+                Log::info('No calendar ID found for band!, skipping calendar deletion for booking ID: ' . $this->booking->id);
+                return;
+            }
+
+            $calendarService = new CalendarService($this->booking->band, 'booking');
             $calendarService->deleteBookingFromCalendar($this->booking);
         } catch (\Exception $e) {
             Log::error('Failed to delete booking from calendar in observer: ' . $e->getMessage());
